Update 'create_right_cadastre_wizard' wizard
